+++ src/app/fallback/synchronization/synchronization.component.ts
@@ -1,0 +1,16 @@
+import { Component } from "@angular/core";
+import {FooterToolbarService} from "../../../services/footer-toolbar.service";
+import {InAppBrowser} from "@ionic-native/in-app-browser";
+
+@Component({
+  templateUrl: "synchronization.html"
+})
+export class SynchronizationPage {
+
+  constructor(
+      readonly footerToolbar: FooterToolbarService
+  ) {
+
+  }
+
+}