+++ src/presenters/course-presenter.ts
@@ -3,7 +3,7 @@ import {GenericILIASObjectPresenter} from "./object-presenter";
 export class CourseObjectPresenter extends GenericILIASObjectPresenter {
 
     icon(): string {
-        return "./assets/icon/icon_crs.svg";
+        return "assets/icon/obj_course.svg";
     }
 
     showTypeAsText(): boolean {