+++ src/presenters/folder-presenter.ts
@@ -3,7 +3,7 @@ import {GenericILIASObjectPresenter} from "./object-presenter";
 export class FolderObjectPresenter extends GenericILIASObjectPresenter {
 
     icon(): string {
-        return "./assets/icon/icon_fold.svg";
+        return "assets/icon/obj_folder.svg";
     }
 
     showTypeAsText(): boolean {