+++ CHANGELOG.md
@@ -1,6 +1,43 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] – 2019-06-19
+
+### Changed
+- Synchronization revised
+- User Interface completely revised
+- File Handling revised
+- Faster loading of metadata
+- Live loading of metadata
+- Favorites are offline available now
+
+### Added
+- Branding provider
+- Branding folder
+- Branded App files
+- Tile Home screen
+- Access to the repository (join courses)
+
+## [2.0.2] – 2018-10-02
+
+### Fixed
+- OnBoarding ( A bug that caused certain Android smartphones, especially Huawei & Samsung, that couldn't Skip onboarding)
+- Loading spinner
+
+### Changed
+- Updated GoogleMaps
+- Updated DocumentHandler for iOS
+
+### Added
+- Support for Android Oreo
+- Support for iOS 12
+- User Information on intial Sync
+
+## [2.0.1] – 2018-07-08
+### Added
+- OnBoarding
+- Information concerning EU DSGVO / GDPR   
+
 ## [2.0.0] – 2018-04-13
 ### Added
 - Learnplace feature